# docs/swagger.yaml
definitions:
  lib.ErrResponse:
    properties:
      data: {}
      error:
        type: string
      ok:
        type: boolean
    type: object
  lib.Response:
    properties:
      data: {}
      ok:
        type: boolean
    type: object
  models.CreateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        minimum: 0
        type: integer
      service_name:
        minLength: 1
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  models.SubscriptionListResponse:
    properties:
      limit:
        type: integer
      next_page:
        type: string
      page:
        type: integer
      prev_page:
        type: string
      subscriptions:
        items:
          $ref: '#/definitions/models.SubscriptionResponse'
        type: array
      total:
        type: integer
    type: object
  models.SubscriptionResponse:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
info:
  contact: {}
  termsOfService: http://swagger.io/terms/
  title: Subscription Service API
paths:
  /subscriptions:
    delete:
      parameters:
      - description: Subscription id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.ErrResponse'
      summary: Delete subscription by id
    get:
      parameters:
      - description: Subscription id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SubscriptionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.ErrResponse'
      summary: Get subscription by id
    patch:
      consumes:
      - application/json
      parameters:
      - description: Subscription id
        in: query
        name: id
        required: true
        type: string
      - description: Subscription info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SubscriptionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.ErrResponse'
      summary: Update subscription
    post:
      consumes:
      - application/json
      parameters:
      - description: Subscription info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SubscriptionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.ErrResponse'
      summary: Create subscription
  /subscriptions/list:
    get:
      parameters:
      - description: Page of list
        in: query
        name: page
        type: integer
      - description: Limit of list's page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SubscriptionListResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.ErrResponse'
      summary: Get a list of subscriptions
  /subscriptions/total-cost:
    get:
      parameters:
      - description: Subscription start date
        in: query
        name: start_date
        required: true
        type: string
      - description: Subscription end date
        in: query
        name: end_date
        required: true
        type: string
      - description: Subscription service name
        in: query
        name: service_name
        type: string
      - description: Subscription user id
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                data:
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.ErrResponse'
      summary: Сalculate the total cost of subscriptions for a specified period with
        the ability to filter by user or service
swagger: "2.0"